session-agent/requirements.txt
fastapi
uvicorn
pydantic
langchain
openai
python-dotenv
motor